Remove python/michelangelo/canvas/ and relocate JSONData to lib/shared
Port the live JSONData module (json_data.py, field.py, and tests) from
canvas/lib/shared/json_data/ to lib/shared/json_data/, matching the
existing lib/shared/utils/ submodule convention. Update the one external
consumer (uniflow/registration/config_builder.py) to import from the
new path.

Delete the remaining dead code in canvas/:
- canvas/lib/model/tte/ (TTE model SDK, zero external consumers)
- canvas/schema/v2alpha1/ (config/job_specs, zero external consumers)
- canvas/lib/shared/utils.py (only used by deleted TTE code)
- All associated test files and binary test fixtures
