--- python/michelangelo/canvas/lib/model/tte/unit_test_model_files/config.json
@@ -1,21 +1,0 @@
-{
-  "attention_probs_dropout_prob": 0.1,
-  "classifier_dropout": null,
-  "gradient_checkpointing": false,
-  "hidden_act": "gelu",
-  "hidden_dropout_prob": 0.1,
-  "hidden_size": 32,
-  "initializer_range": 0.02,
-  "intermediate_size": 37,
-  "layer_norm_eps": 1e-12,
-  "max_position_embeddings": 512,
-  "model_type": "bert",
-  "num_attention_heads": 4,
-  "num_hidden_layers": 5,
-  "pad_token_id": 0,
-  "position_embedding_type": "absolute",
-  "transformers_version": "4.11.0.dev0",
-  "type_vocab_size": 16,
-  "use_cache": true,
-  "vocab_size": 1124
-}
