--- python/michelangelo/canvas/lib/model/tte/unit_test_model_files/special_tokens_map.json
@@ -1,1 +1,0 @@
-{"unk_token": "[UNK]", "sep_token": "[SEP]", "pad_token": "[PAD]", "cls_token": "[CLS]", "mask_token": "[MASK]"}
