--- python/michelangelo/canvas/lib/model/tte/unit_test_model_files/vocab.txt
@@ -1,1124 +1,0 @@
-[PAD]
-[UNK]
-[CLS]
-[SEP]
-[MASK]
-!
-"
-#
-$
-%
-&
-'
-(
-)
-*
-+
-,
--
-.
-/
-0
-1
-2
-3
-4
-5
-6
-7
-8
-9
-:
-;
-<
-=
->
-?
-@
-[
-\
-]
-^
-_
-`
-a
-b
-c
-d
-e
-f
-g
-h
-i
-j
-k
-l
-m
-n
-o
-p
-q
-r
-s
-t
-u
-v
-w
-x
-y
-z
-|
-}
-~
-¡
-¢
-£
-¥
-§
-°
-±
-²
-³
-´
-µ
-·
-º
-½
-¿
-×
-ß
-æ
-ð
-ø
-þ
-đ
-ħ
-ı
-ł
-œ
-ɐ
-ɑ
-ɒ
-ɔ
-ə
-ɛ
-ɜ
-ɡ
-ɢ
-ɪ
-ɫ
-ɳ
-ɽ
-ɾ
-ʁ
-ʃ
-ʊ
-ʋ
-ʒ
-ʔ
-ʕ
-ʲ
-ʻ
-ʼ
-ʾ
-ʿ
-ˈ
-ˌ
-ː
-α
-β
-γ
-δ
-ε
-η
-θ
-ι
-κ
-λ
-μ
-ν
-ξ
-ο
-π
-ρ
-ς
-σ
-τ
-υ
-φ
-χ
-ψ
-ω
-а
-б
-в
-г
-д
-е
-ж
-з
-и
-к
-л
-м
-н
-о
-п
-р
-с
-т
-у
-х
-ц
-ш
-ъ
-ы
-ь
-ю
-я
-є
-א
-ב
-ג
-ה
-ו
-ז
-ח
-י
-ל
-ם
-מ
-ן
-נ
-ס
-ף
-פ
-צ
-ר
-ש
-ת
-ء
-ا
-ب
-ة
-ت
-ث
-ج
-ح
-خ
-د
-ذ
-ر
-س
-ش
-ص
-ع
-ف
-ق
-ك
-ل
-م
-ن
-ه
-و
-ي
-ܐ
-ܕ
-ܗ
-ܝ
-ܠ
-ܢ
-ܬ
-अ
-ई
-क
-ग
-ण
-त
-द
-न
-प
-ब
-म
-य
-र
-ल
-व
-स
-ह
-ा
-ि
-আ
-ল
-হ
-া
-ਅ
-ਲ
-ਹ
-ਾ
-അ
-ള
-ഹ
-ാ
-ก
-ค
-ง
-ช
-ซ
-ญ
-ฐ
-ณ
-ด
-ต
-น
-บ
-ป
-พ
-ภ
-ม
-ย
-ร
-ล
-ว
-ศ
-ษ
-ส
-ห
-อ
-ฮ
-ะ
-า
-เ
-แ
-ไ
-ა
-ბ
-გ
-დ
-ე
-ვ
-ზ
-თ
-ი
-კ
-ლ
-მ
-ნ
-ო
-პ
-ჟ
-რ
-ს
-ტ
-უ
-ფ
-ქ
-ღ
-ყ
-შ
-ჩ
-ც
-ძ
-წ
-ჭ
-ხ
-ჯ
-ჰ
-ჱ
-ჲ
-ჳ
-ჴ
-ჵ
-ჶ
-ჷ
-ჸ
-ჹ
-ჺ
-჻
-ᄃ
-ᄅ
-ᄇ
-ᄋ
-ᄌ
-ᅡ
-ᅢ
-ᅦ
-ᅧ
-ᅩ
-ᅮ
-ᅵ
-ᆨ
-ᆫ
-ᆯ
-ᆸ
-ᆼ
-ᵻ
-‐
-‑
-–
-—
-―
-‘
-’
-“
-”
-„
-†
-‡
-•
-…
-′
-″
-⁄
-₣
-₤
-€
-₹
-⅓
-⅔
-→
-−
-≡
-≤
-①
-☉
-☫
-♀
-♭
-♯
-⚳
-ⴀ
-ⴂ
-ⴃ
-ⴈ
-ⴌ
-ⴕ
-ⴟ
-〈
-〉
-〜
-あ
-い
-う
-お
-か
-き
-く
-け
-こ
-さ
-し
-す
-せ
-た
-ち
-っ
-つ
-と
-な
-に
-の
-は
-ひ
-ふ
-ほ
-ま
-み
-め
-も
-ゃ
-ゆ
-ょ
-ら
-り
-る
-れ
-わ
-を
-ん
-ァ
-ア
-ィ
-イ
-ゥ
-ウ
-ェ
-エ
-ォ
-オ
-カ
-キ
-ク
-ケ
-コ
-サ
-シ
-ス
-セ
-タ
-チ
-ッ
-ツ
-テ
-ト
-ナ
-ニ
-ネ
-ノ
-ハ
-フ
-ヘ
-マ
-ミ
-ム
-モ
-ャ
-ュ
-ョ
-ラ
-リ
-ル
-レ
-ロ
-ン
-・
-ー
-一
-七
-下
-世
-丙
-中
-主
-乃
-之
-乙
-九
-二
-云
-人
-今
-付
-作
-侗
-依
-信
-傳
-儚
-充
-光
-全
-兵
-其
-具
-円
-再
-出
-判
-前
-剛
-劇
-劉
-動
-化
-北
-华
-厂
-去
-古
-可
-台
-史
-同
-名
-君
-吳
-周
-命
-和
-咲
-善
-四
-國
-園
-圣
-在
-坂
-堤
-場
-塘
-夕
-大
-天
-夫
-女
-妙
-姚
-子
-孟
-守
-安
-宋
-完
-宗
-宝
-宫
-寝
-寺
-小
-少
-尾
-山
-岳
-川
-州
-巳
-市
-師
-平
-广
-庆
-府
-座
-廬
-建
-式
-張
-彌
-彩
-彼
-後
-御
-德
-思
-愛
-憑
-憶
-應
-懷
-战
-戦
-扈
-技
-拉
-拳
-挑
-揺
-攻
-放
-政
-散
-斯
-方
-日
-旦
-旭
-昌
-明
-星
-春
-晋
-景
-曦
-月
-望
-未
-本
-李
-村
-杜
-束
-来
-林
-桜
-梶
-棘
-椎
-楊
-楚
-榮
-橘
-機
-正
-殻
-殿
-母
-水
-汉
-沂
-沙
-河
-泗
-波
-泣
-洪
-淹
-清
-湯
-漢
-澄
-澤
-火
-灯
-灵
-灼
-焼
-熱
-物
-狐
-狸
-玄
-王
-玩
-珂
-珙
-球
-理
-琦
-琪
-瓊
-生
-田
-畢
-番
-瘡
-白
-皮
-真
-砲
-礮
-祈
-神
-祠
-秋
-空
-立
-精
-約
-絵
-織
-義
-翠
-者
-耕
-肖
-胡
-膀
-臂
-興
-良
-花
-芳
-芽
-若
-英
-藕
-藥
-蘄
-蘇
-行
-裁
-規
-覺
-观
-解
-記
-誓
-誡
-誰
-謎
-许
-谭
-豪
-豫
-費
-贵
-赤
-趙
-足
-跡
-転
-辛
-逆
-遇
-運
-過
-遠
-選
-邦
-邱
-部
-郭
-都
-酈
-里
-野
-金
-銃
-鋼
-錄
-錡
-鍵
-鐵
-钱
-铁
-關
-防
-阿
-陈
-陳
-陽
-隊
-階
-集
-雪
-雲
-霖
-霹
-靂
-韓
-願
-顯
-颜
-马
-高
-龍
-ﷲ
-ﻋ
-／
-３
-～
-##e
-##t
-##f
-##r
-##n
-##o
-##a
-##3
-##4
-##0
-##i
-##u
-##b
-##s
-##g
-##p
-##l
-##d
-##c
-##v
-##y
-##h
-##k
-##ˈ
-##m
-##9
-##1
-##w
-##2
-##x
-##z
-##β
-##ø
-##7
-##5
-##j
-##8
-##ł
-##q
-##6
-##は
-##り
-##α
-##ο
-##ς
-##く
-##ん
-##ə
-##ɔ
-##ʔ
-##о
-##в
-##е
-##т
-##с
-##к
-##а
-##я
-##め
-##ي
-##د
-##ا
-##ر
-##ा
-##ई
-##ル
-##ハ
-##ト
-##ス
-##ψ
-##ι
-##セ
-##リ
-##オ
-##ロ
-##π
-##τ
-##κ
-##ν
-##θ
-##η
-##μ
-##ッ
-##フ
-##ォ
-##ン
-##ア
-##ク
-##タ
-##ー
-##д
-##р
-##л
-##こ
-##と
-##ı
-##シ
-##ʻ
-##ი
-##ρ
-##ן
-##ण
-##न
-##क
-##ю
-##з
-##ラ
-##ュ
-##な
-##き
-##س
-##م
-##ت
-##ნ
-##ლ
-##მ
-##წ
-##ფ
-##ე
-##ː
-##ゆ
-##ل
-##ب
-##ن
-##ニ
-##ל
-##ה
-##い
-##ち
-##ゃ
-##י
-##פ
-##た
-##ه
-##ხ
-##ჳ
-##ს
-##ß
-##ʒ
-##æ
-##ล
-##อ
-##ฮ
-##あ
-##ɒ
-##ᵻ
-##ʊ
-##ס
-##ð
-##ო
-##თ
-##ა
-##ვ
-##რ
-##უ
-##ェ
-##テ
-##ィ
-##ˌ
-##イ
-##ف
-##ტ
-##ء
-##х
-##แ
-##ก
-##γ
-##つ
-##か
-##エ
-##đ
-##ك
-##и
-##н
-##せ
-##マ
-##ɪ
-##δ
-##გ
-##す
-##ɛ
-##ق
-##ع
-##の
-##ы
-##ल
-##ゥ
-##ナ
-##λ
-##χ
-##ਲ
-##ਾ
-##ਹ
-##ε
-##ъ
-##п
-##ו
-##ף
-##ɡ
-##و
-##ح
-##う
-##द
-##ɫ
-##ค
-##ร
-##ว
-##ต
-##レ
-##チ
-##σ
-##υ
-##カ
-##ป
-##ย
-##ャ
-##モ
-##ω
-##ɑ
-##ま
-##る
-##พ
-##у
-##ث
-##ミ
-##っ
-##し
-##ょ
-##に
-##れ
-##დ
-##ʿ
-##ム
-##र
-##ง
-##ไ
-##ช
-##า
-##є
-##स
-##ם
-##ة
-##ള
-##も
-##ჲ
-##わ
-##サ
-##א
-##ב
-##ר
-##ল
-##お
-##ɐ
-##ش
-##ภ
-##ด
-##ひ
-##ც
-##ɾ
-##ʲ
-##ウ
-##ァ
-##キ
-##３
-##を
-##ᅡ
-##ᆯ
-##ᄅ
-##ʃ
-##ჱ
-##º
-##ɳ
-##ら
-##ж
-##خ
-##ʁ
-##ะ
-##น
-##ษ
-##ฐ
-##м
-##ह
-##া
-##হ
-##व
-##œ
-##ᄇ
-##ᅮ
-##ᄃ
-##ᅢ
-##ɜ
-##ि
-##ग
-##ξ
-##נ
-##ქ
-##ص
-##ാ
-##ഹ
-##г
-##ܠ
-##ܗ
-##ܐ
-##ɽ
-##ツ
-##ซ
-##ш
-##ц
-##ь
-##ネ
-##ョ
-##ณ
-##ม
-##ノ
-##ש
-##מ
-##צ
-##ᅦ
-##ᅩ
-##ᆨ
-##ᅵ
-##ᆸ
-##ᅧ
-##ᆼ
-##ᄋ
-##ᆫ
-##เ
-##ذ
-##ܕ
-##ܝ
-##ܢ
-##ܬ
-##み
-##ת
