--- python/michelangelo/canvas/lib/shared/json_data/__init__.py
@@ -1,4 +1,0 @@
-from .json_data import JSONData
-from .field import field, one_of
-
-__all__ = ["JSONData", "field", "one_of"]
