--- python/michelangelo/canvas/lib/shared/json_data/tests/__init__.py
@@ -1,1 +1,0 @@
-# Test package for JSONData
