+++ python/michelangelo/lib/shared/json_data/json_data.py
@@ -1,0 +1,231 @@
+"""JSONData base model with constrained pydantic features for uniflow."""
+
+import typing
+from enum import Enum
+
+import pydantic
+from pydantic import BaseModel, ConfigDict, SerializationInfo, model_serializer
+from pydantic.fields import FieldInfo
+from pydantic_core import PydanticUndefined
+
+from michelangelo.lib.shared.json_data.field import _OneOf, field
+
+
+class JSONData(BaseModel):
+    """Base class for structured configuration classes.
+
+    Subclass of ``pydantic.BaseModel`` that:
+
+    1. Adds default values for fields based on their types.
+    2. Adds ``one_of`` mutual-exclusion constraints.
+    3. Restricts pydantic features to what uniflow's Go service
+       and Starlark can safely support.
+    """
+
+    def __init_subclass__(cls, **kwargs):
+        """Validate field types and register one-of constraints."""
+        for base in cls.__bases__:
+            if JSONData != base and not issubclass(base, JSONData):
+                raise TypeError(
+                    f"{base.__name__} is not a subclass of"
+                    f" {JSONData.__name__}. {JSONData.__name__}"
+                    f" classes can only inherit from"
+                    f" {JSONData.__name__} or its subclasses."
+                )
+
+        fields = {}
+        annotations = typing.get_type_hints(cls)
+        for field_name, field_type in annotations.items():
+            if field_name.startswith("_"):
+                continue
+            type_info = _get_type_info(field_name, field_type)
+
+            if field_name in cls.__dict__:
+                v = getattr(cls, field_name)
+                field_info = v if isinstance(v, FieldInfo) else field(default=v)
+            else:
+                field_info = field()
+
+            if (
+                field_info.default is PydanticUndefined
+                and not field_info.json_schema_extra["json_data_field"].get("required")
+            ):
+                field_info.default = _get_default_value(type_info, field_type)
+            setattr(cls, field_name, field_info)
+            field_info.json_schema_extra["json_data_field"] |= type_info
+            fields[field_name] = (field_type, field_info)
+
+        json_data_info = {}
+        for attr_name, v in cls.__private_attributes__.items():
+            if isinstance(v.default, _OneOf):
+                oneof = v.default
+                for f in oneof.fields:
+                    if f not in fields:
+                        raise ValueError(
+                            f"Field in one_of '{attr_name}' does"
+                            f" not exist. No field named '{f}'"
+                            f" in class {cls.__name__}."
+                        )
+                    f_info = fields[f][1]
+                    if not f_info.json_schema_extra["json_data_field"].get("nullable"):
+                        raise TypeError(
+                            f"Field '{f}' in one_of"
+                            f" '{attr_name}' is not optional."
+                            " All the fields in oneof must be"
+                            " optional (nullable)."
+                        )
+                one_of_list = json_data_info.get("oneof", [])
+                one_of_list.append(oneof.model_dump())
+                json_data_info["oneof"] = one_of_list
+
+        json_schema_extra = {"json_data_object": json_data_info}
+        cls.model_config = ConfigDict(json_schema_extra=json_schema_extra)
+        cls.__validate__ = pydantic.model_validator(mode="after")(_validate_model_)
+
+    @model_serializer(mode="wrap")
+    def serialize_model(
+        self, handler, info: SerializationInfo
+    ) -> dict[str, typing.Any]:
+        """Serialize with optional UniflowCodec metadata."""
+        dump = handler(self, info)
+        if info.context and info.context.get("UniflowCodec", False):
+            dump |= {
+                "__codec__": "dataclass",
+                "__class__": (f"{type(self).__module__}.{type(self).__name__}"),
+            }
+        return dump
+
+
+def _get_type_info(
+    field_name: str, field_type: type, position: str = "Field"
+) -> dict[str, typing.Any]:
+    origin = typing.get_origin(field_type)
+
+    if origin is typing.Union:
+        if position in ["List item", "Dict value"]:
+            raise TypeError(
+                f"{position} type '{field_type}' is not"
+                f" supported in JSONData class."
+                f" Field: '{field_name}'"
+            )
+        type_list = typing.get_args(field_type)
+        if len(type_list) > 2 or type(None) not in type_list:
+            raise TypeError(
+                f"Field type '{field_type}' is not supported"
+                f" in JSONData class. Field: '{field_name}'"
+            )
+        type_info = _get_type_info(
+            field_name,
+            next(iter(t for t in type_list if t is not type(None))),
+        )
+        type_info["nullable"] = True
+        return type_info
+
+    if field_type in [bool, int, float, str]:
+        return {"type": field_type.__name__}
+
+    if isinstance(field_type, type) and issubclass(field_type, Enum):
+        if issubclass(field_type, str):
+            return {"type": field_type.__name__}
+        raise TypeError(
+            f"Enum type {field_type} is not supported in"
+            f" JSONData class. Only string Enum is supported."
+            f" Field: '{field_name}'"
+        )
+
+    if (
+        origin is None
+        and isinstance(field_type, type)
+        and issubclass(field_type, JSONData)
+    ):
+        return {"type": field_type.__name__}
+
+    if field_type is list or origin is list:
+        type_args = typing.get_args(field_type)
+        if len(type_args) == 0:
+            item_type = typing.Any
+        elif len(type_args) == 1:
+            item_type = typing.get_args(field_type)[0]
+        else:
+            raise TypeError(f"Invalid list type: {field_type}. Field: '{field_name}'")
+
+        return {
+            "type": "list",
+            "item_type": _get_type_info(field_name, item_type, "List item"),
+        }
+
+    if field_type is dict or origin is dict:
+        type_args = typing.get_args(field_type)
+        if len(type_args) == 2:
+            key_type = type_args[0]
+            value_type = type_args[1]
+        elif len(type_args) == 0:
+            key_type = str
+            value_type = typing.Any
+        else:
+            raise TypeError(
+                f"Invalid dictionary type: {field_type}. Field: '{field_name}'"
+            )
+
+        if key_type is not str:
+            raise TypeError(
+                f"Invalid dictionary type: {field_type}."
+                " Dictionary keys must be strings."
+                f" Field: '{field_name}'"
+            )
+
+        return {
+            "type": "map",
+            "key_type": {"type": "string"},
+            "value_type": _get_type_info(field_name, value_type, "Dict value"),
+        }
+
+    if field_type is typing.Any:
+        return {"type": "any"}
+
+    raise TypeError(
+        f"{position} type {field_type} is not supported in"
+        f" JSONData class. Field: '{field_name}'"
+    )
+
+
+def _get_default_value(type_info: dict[str, str], field_type: type) -> typing.Any:
+    if type_info.get("nullable"):
+        return None
+
+    t = type_info["type"]
+    defaults = {"bool": False, "int": 0, "float": 0.0, "str": ""}
+    if t in defaults:
+        return defaults[t]
+
+    origin = typing.get_origin(field_type)
+    if origin is None and isinstance(field_type, type):
+        if issubclass(field_type, JSONData):
+            return field_type()
+        if issubclass(field_type, Enum):
+            return next(iter(field_type))
+
+    if field_type is list or origin is list:
+        return []
+    if field_type is dict or origin is dict:
+        return {}
+
+    return None
+
+
+def _validate_model_(self: BaseModel):
+    """Run json_data-specific validations (oneof constraints)."""
+    one_of_list = self.model_config["json_schema_extra"]["json_data_object"].get(
+        "oneof", []
+    )
+    for i in one_of_list:
+        one_of = _OneOf(**i)
+        set_fields = [f for f in one_of.fields if getattr(self, f) is not None]
+        if len(set_fields) == 0 and one_of.required:
+            raise ValueError(f"One field in {one_of.fields} must be set (not None).")
+        if len(set_fields) > 1:
+            raise ValueError(
+                f"More than one field in {one_of.fields}"
+                f" are set (not None): {set_fields}."
+            )
+    return self
