+++ python/michelangelo/lib/shared/json_data/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Test package for JSONData."""
